Compare result cards by id
When testing a pulse, only the id is POSTed, not the name. The code
was previously using name to connect the card associated with the
result to the card requested in the Pulse. This works fine for timed
pulse sends as the information is pulled from the database, but for
tests, it's only the data in the POST payload (which only includes the
id). Made the switch to ids so it works in both cases.